stable 3 march

+++ controllers/LessonController.js
@@ -1,5 +1,5 @@
 const  Lesson  = require('../models/Lessons');
-
+const {Progress}=require('../models/Progresses')
 // Создать урок
 exports.createLesson = async (req, res) => {
   console.log('Create Lessaon started!',req.body);
@@ -71,14 +71,32 @@ exports.updateLesson = async (req, res) => {
 // Удалить урок
 exports.deleteLesson = async (req, res) => {
   try {
-    const deleted = await Lesson.destroy({
-      where: { id: req.params.id },
+    
+    const id = req.params.id;
+    console.log('lessonId= ',id)
+
+    // Находим урок
+    const lesson = await Lesson.findByPk(id);
+    if (!lesson) {
+      return res.status(404).json({ error: 'Lesson not found' });
+    }
+
+    // Удаляем связанные записи в progresses (если не используете CASCADE)
+    await Progress.destroy({
+      where: { lesson_id: id },
     });
+
+    // Удаляем урок
+    const deleted = await Lesson.destroy({
+      where: { id: id },
+    });
+
     if (deleted) {
       return res.status(204).send();
     }
     res.status(404).json({ error: 'Lesson not found' });
   } catch (error) {
+    console.error('Ошибка при удалении урока:', error);
     res.status(500).json({ error: error.message });
   }
 };
